backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "faster-whisper>=1.2.1",
    "kokoro-onnx>=0.5.0",
    "livekit-agents>=1.4.5",
    "livekit-plugins-openai>=1.4.5",
    "livekit-plugins-silero>=1.4.5",
    "livekit-plugins-noise-cancellation~=0.2",
    "llama-cpp-python>=0.3.16",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "soundfile>=0.13.1",
    "rich>=14.3.3",
]
